+++ src/client/js/dateViewUpdater.js
@@ -1,5 +1,3 @@
-import airplane from '../../assets/images/airplane.svg';
-import calendar from '../../assets/images/calendar.svg';
 const dateViewUpdater = (function () {
     let _applicationState;
 
@@ -168,35 +166,6 @@ const dateViewUpdater = (function () {
         return day < 10 ? '0' + day : day;
     }
 
-    function updateDateView() {
-        const startDate = new Date(_applicationState.startDate);
-        const endDate = new Date(_applicationState.endDate);
-        const options = {
-            year: 'numeric',
-            month: 'long',
-            day: 'numeric',
-        };
-        document.getElementById(
-            'newTrip__startDate'
-        ).textContent = startDate.toLocaleDateString('en-US', options);
-        document.getElementById(
-            'newTrip__endDate'
-        ).textContent = endDate.toLocaleDateString('en-US', options);
-        document.getElementById('newTrip__daysToTrip').textContent =
-            _applicationState.daysToTrip == 0
-                ? 'today! Get Ready!'
-                : 'in ' +
-                  _applicationState.daysToTrip +
-                  ' day' +
-                  (_applicationState.daysToTrip != 1 ? 's' : '');
-        document.getElementById('newTrip__duration').textContent =
-            _applicationState.duration +
-            ' day' +
-            (_applicationState.duration != 1 ? 's' : '');
-        document.querySelector('.time__icon--calendar').innerHTML = calendar;
-        document.querySelector('.time__icon--airplane').innerHTML = airplane;
-    }
-
     function showDateError(id, message) {
         const errorContainer = document.getElementById(
             'tripSelection__error--' + id
@@ -223,7 +192,6 @@ const dateViewUpdater = (function () {
         renderFallbackDateInput,
         populateYears,
         populateDays,
-        updateDateView,
         showDateError,
         clearDateErrors,
     };
