+++ src/client/js/viewUpdater.js
@@ -1,3 +1,5 @@
+import airplane from '../../assets/images/airplane.svg';
+import calendar from '../../assets/images/calendar.svg';
 import { dateViewUpdater } from './dateViewUpdater';
 import { cityViewUpdater } from './cityViewUpdater';
 import { pictureViewUpdater } from './pictureViewUpdater';
@@ -14,6 +16,36 @@ const viewUpdater = (function () {
             _applicationState.province +
             ', ' +
             _applicationState.country.toUpperCase();
+        updateTimeElement();
+    }
+
+    function updateTimeElement() {
+        const startDate = new Date(_applicationState.startDate);
+        const endDate = new Date(_applicationState.endDate);
+        const options = {
+            year: 'numeric',
+            month: 'long',
+            day: 'numeric',
+        };
+        document.getElementById(
+            'newTrip__startDate'
+        ).textContent = startDate.toLocaleDateString('en-US', options);
+        document.getElementById(
+            'newTrip__endDate'
+        ).textContent = endDate.toLocaleDateString('en-US', options);
+        document.getElementById('newTrip__daysToTrip').textContent =
+            _applicationState.daysToTrip == 0
+                ? 'today! Get Ready!'
+                : 'in ' +
+                  _applicationState.daysToTrip +
+                  ' day' +
+                  (_applicationState.daysToTrip != 1 ? 's' : '');
+        document.getElementById('newTrip__duration').textContent =
+            _applicationState.duration +
+            ' day' +
+            (_applicationState.duration != 1 ? 's' : '');
+        document.querySelector('.time__icon--calendar').innerHTML = calendar;
+        document.querySelector('.time__icon--airplane').innerHTML = airplane;
     }
 
     // Forwarding commands to dateViewUpdater
@@ -31,10 +63,6 @@ const viewUpdater = (function () {
 
     function populateDays(id, previousDay) {
         dateViewUpdater.populateDays(id, previousDay);
-    }
-
-    function updateDateView() {
-        dateViewUpdater.updateDateView();
     }
 
     function showDateError(id, message) {
@@ -98,7 +126,6 @@ const viewUpdater = (function () {
         renderFallbackDateInput,
         populateYears,
         populateDays,
-        updateDateView,
         showDateError,
         clearDateErrors,
         createNewCityList,
